test(replay): Properly split tests between integration & unit tests
