+++ packages/replay/test/integration/coreHandlers/handleScope.test.ts
