+++ packages/replay/test/integration/integrationSettings.test.ts
@@ -1,5 +1,5 @@
 import { MASK_ALL_TEXT_SELECTOR } from '../../src/constants';
-import { mockSdk } from './../index';
+import { mockSdk } from '../index';
 
 describe('integration settings', () => {
   beforeEach(() => {
