+++ packages/replay/test/integration/sampling.test.ts
@@ -1,6 +1,6 @@
 // mock functions need to be imported first
-import { mockRrweb, mockSdk } from './../index';
-import { useFakeTimers } from './../utils/use-fake-timers';
+import { mockRrweb, mockSdk } from '../index';
+import { useFakeTimers } from '../utils/use-fake-timers';
 
 useFakeTimers();
